[refactor] rename Stage to Map and Monitor
