--- examples/worldview/src/features/display/index.js
@@ -1,3 +1,0 @@
-export {DisplayContainer as Display} from './DisplayContainer';
-export {default as displayReducer, updateViewState, viewStateSelector} from './displaySlice';
-export {useWhenReady} from './hooks';
