+++ examples/worldview/src/features/map/hooks.js
@@ -1,7 +1,7 @@
 import {useEffect} from 'react';
 import {useSelector} from 'react-redux';
 export const useWhenReady = onReady => {
-  const ready = useSelector(state => state.hubbleGl.display.ready);
+  const ready = useSelector(state => state.hubbleGl.map.ready);
   useEffect(() => {
     if (ready) onReady();
   }, [ready]);
