+++ examples/worldview/src/features/map/index.js
@@ -1,0 +1,3 @@
+export {MapContainer as Map} from './MapContainer';
+export {default as mapReducer, updateViewState, viewStateSelector} from './mapSlice';
+export {useWhenReady} from './hooks';
