+++ examples/worldview/src/features/map/store/render-settings.js
