+++ examples/worldview/src/features/monitor/Icons.js
